# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "description": "Returns the complete list of registered users. Requires admin or superuser role.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Admin/superuser creates a user directly. For public registration, use POST /users/register instead.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation failed or duplicate email"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Create a new user (admin-only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/search": {
      "get": {
        "operationId": "UsersController_search",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Search users by name, email, or role",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/pending": {
      "get": {
        "description": "Returns users awaiting approval (status=pending)",
        "operationId": "UsersController_getPending",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending users retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get pending users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/approved": {
      "get": {
        "operationId": "UsersController_getApproved",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved users retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get approved users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/rejected": {
      "get": {
        "operationId": "UsersController_getRejected",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rejected users retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get rejected users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Permanently removes a user from the system. Superuser only.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/register": {
      "post": {
        "description": "Anyone can register. Donors and beneficiaries get instant approval. Volunteers and admins go to pending status.",
        "operationId": "UsersController_register",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Validation failed or duplicate email"
          }
        },
        "summary": "Public registration (no role header required)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/signin": {
      "post": {
        "description": "Authenticates a user with email, password, and role. Returns user data on success.",
        "operationId": "UsersController_signIn",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sign-in failed — invalid credentials or account not approved\n\nSigned in successfully"
          }
        },
        "summary": "Sign in a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/approve": {
      "patch": {
        "description": "Changes user status from pending to approved. Superuser only.",
        "operationId": "UsersController_approve",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User approved successfully"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Approve a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/reject": {
      "patch": {
        "description": "Changes user status to rejected. Superuser only.",
        "operationId": "UsersController_reject",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC (superuser, admin, volunteer, donor, beneficiary)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User rejected"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Reject a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindred Platform API",
    "description": "Backend REST API for Kindred — India's Volunteer Platform. Supports Users, Organizations, Volunteers, Programs, Requests, Tasks, Hour Logs, Resource Donations, Program Assignments, Program Applications, and Subtasks. Role-based access control via the \"role\" header.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Aarav Mehta"
          },
          "email": {
            "type": "string",
            "description": "Valid email address",
            "example": "aarav.mehta@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Account password (min 6 chars)",
            "example": "Secure@456"
          },
          "role": {
            "type": "string",
            "description": "User role on the platform",
            "example": "volunteer",
            "enum": [
              "superuser",
              "admin",
              "volunteer",
              "donor",
              "beneficiary"
            ]
          },
          "org": {
            "type": "string",
            "description": "Organization name (required for admin role)",
            "example": "GreenFuture Foundation"
          },
          "phone": {
            "type": "string",
            "description": "Phone number with country code",
            "example": "+91 98765 43210"
          },
          "bio": {
            "type": "string",
            "description": "Short bio or introduction",
            "example": "Passionate about rural education and healthcare access"
          },
          "skills": {
            "description": "List of skills",
            "example": [
              "Teaching",
              "First Aid",
              "Logistics"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Aarav Mehta"
          },
          "email": {
            "type": "string",
            "description": "Valid email address",
            "example": "aarav.mehta@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Account password (min 6 chars)",
            "example": "Secure@456"
          },
          "role": {
            "type": "string",
            "description": "User role on the platform",
            "example": "volunteer",
            "enum": [
              "superuser",
              "admin",
              "volunteer",
              "donor",
              "beneficiary"
            ]
          },
          "org": {
            "type": "string",
            "description": "Organization name (required for admin role)",
            "example": "GreenFuture Foundation"
          },
          "phone": {
            "type": "string",
            "description": "Phone number with country code",
            "example": "+91 98765 43210"
          },
          "bio": {
            "type": "string",
            "description": "Short bio or introduction",
            "example": "Passionate about rural education and healthcare access"
          },
          "skills": {
            "description": "List of skills",
            "example": [
              "Teaching",
              "First Aid",
              "Logistics"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "example": "approved",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          }
        }
      },
...
    }
  }
}